* Create development branch and custom Docker image
* Update order of operations in cron script
* Use staging path for GCS outputs in other branches

+++ src/scripts/cron-update.sh
@@ -1,4 +1,4 @@
-#!/bin/sh
+#!/bin/bash
 
 # Copyright 2020 Google LLC
 #
@@ -14,44 +14,76 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+#######################################
+# Run data pipelines and upload outputs to GCS
+# Arguments:
+#   Branch to use for the data pipeline execution. Default: "main".
+#   GCS bucket name to upload results to; empty means do not upload. Default: "".
+#######################################
+
+# This is brittle but prevents from continuing in case of failure since we don't want to overwrite
+# files in the server if anything went wrong
 set -xe
 
+# Parse the arguments
+readonly BRANCH="${1:-main}"
+readonly GCS_OUTPUT_BUCKET=$2
+
 # Clone the repo into a temporary directory
-TMPDIR=$(mktemp -d -t opencovid-$(date +%Y-%m-%d-%H-%M-%S)-XXXX)
-git clone https://github.com/open-covid-19/data.git --single-branch -b main "$TMPDIR/opencovid"
+readonly TMPDIR=$(mktemp -d -t opencovid-$(date +%Y-%m-%d-%H-%M-%S)-XXXX)
+git clone https://github.com/open-covid-19/data.git --single-branch -b $BRANCH "$TMPDIR/opencovid"
+
+# Build the Docker image which contains all of our dependencies
+docker build "$TMPDIR/opencovid/src" -t opencovid
 
 # Download the intermediate files into the working directory
+# We always use the production's intermediate files to ensure reproducibility
+# This system seems brittle, but it lets us pick up the last successful output of any data source
+# as long as it is declared in a pipeline's configuration in case the data source breaks.
+readonly GCS_PROD_BUCKET="covid19-open-data"
 mkdir -p "$TMPDIR/opencovid/output/snapshot"
 mkdir -p "$TMPDIR/opencovid/output/intermediate"
-gsutil -m cp -r gs://covid19-open-data/snapshot "$TMPDIR/opencovid/output/"
-gsutil -m cp -r gs://covid19-open-data/intermediate "$TMPDIR/opencovid/output/"
+gsutil -m cp -r "gs://$GCS_PROD_BUCKET/snapshot" "$TMPDIR/opencovid/output/"
+gsutil -m cp -r "gs://$GCS_PROD_BUCKET/intermediate" "$TMPDIR/opencovid/output/"
 
-# Install dependencies and run the update command in a Docker instance
-docker run -v "$TMPDIR/opencovid":/opencovid -w /opencovid -i python:3.8-buster /bin/bash -s <<EOF
+# Run the update command in a Docker instance using our Docker image
+docker run -v "$TMPDIR/opencovid":/opencovid -w /opencovid -i opencovid:latest /bin/bash -s <<EOF
 cd src
-# Install locales
-apt-get update && apt-get install -yq tzdata locales
-sh -c 'echo en_US.UTF-8 UTF-8 >> /etc/locale.gen'
-sh -c 'echo es_ES.UTF-8 UTF-8 >> /etc/locale.gen'
-locale-gen en_US.UTF-8
-locale-gen es_ES.UTF-8
-dpkg-reconfigure --frontend noninteractive locales
-export LC_ALL="en_US.UTF-8"
-export LC_CTYPE="en_US.UTF-8"
 # Install Python dependencies
-pip install -r requirements.txt
-# Update geography first, since lat-lon values are used in other pipelines
-python3 update.py --only geography
-# Update all the other data pipelines
-python3 update.py --exclude geography
+# Necessary if the deps changed since the Docker image was built
+python3 -m pip install -r requirements.txt
+# Update all the data pipelines
+python3 update.py --no-progress
 # Get the files ready for publishing
 python3 publish.py
 EOF
 
-# Upload the outputs to Google Cloud
-gsutil -m cp -r "$TMPDIR/opencovid/output/snapshot" gs://covid19-open-data/
-gsutil -m cp -r "$TMPDIR/opencovid/output/intermediate" gs://covid19-open-data/
-gsutil -m cp -r "$TMPDIR/opencovid/output/public/v2" gs://covid19-open-data/
+# Known location where the output files are created
+readonly OUTPUT_FOLDER="$TMPDIR/opencovid/output/public/v2"
 
-# Cleanup
-sudo rm -rf $TMPDIR
+# Upload the outputs to Google Cloud Storage
+if [ -z "$GCS_OUTPUT_BUCKET" ]
+then
+    echo "GCS output bucket not set"
+    echo "Output files located in $OUTPUT_FOLDER"
+else
+    if [ $BRANCH == "main" ]
+    then
+        # Outputs from the main branch go into the root folder
+        readonly GCS_OUTPUT_PATH="gs://$GCS_OUTPUT_BUCKET/"
+
+        # Only update the intermediate files if this is for the main branch
+        echo "Uploading intermediate files to GCS bucket $GCS_OUTPUT_BUCKET"
+        gsutil -m cp -r "$TMPDIR/opencovid/output/snapshot" "$GCS_OUTPUT_BUCKET/"
+        gsutil -m cp -r "$TMPDIR/opencovid/output/intermediate" "$GCS_OUTPUT_BUCKET/"
+    else
+        # Outputs from any other branch go into the staging folder
+        readonly GCS_OUTPUT_PATH="gs://$GCS_OUTPUT_BUCKET/staging/$BRANCH/"
+    fi
+
+    echo "Uploading outputs to GCS path $GCS_OUTPUT_PATH"
+    gsutil -m cp -r "$OUTPUT_FOLDER" "$GCS_OUTPUT_PATH"
+
+    # Cleanup needs sudo because Docker creates files using its uid
+    sudo rm -rf $TMPDIR
+fi
